Refactor, add comments
